Organize library modules
