+++ lib/index.js
@@ -5,7 +5,7 @@
 
   // Visualization Libraries
   var libraries = {
-    'default': require('./libraries/default')(Dataviz)
+    'default': require('./libraries')(Dataviz)
     // TODO: add remaining
   };
