+++ package.json
@@ -28,7 +28,7 @@
     "browserify": "^12.0.1",
     "chai": "^3.4.0",
     "chai-dom": "^1.4.0",
-    "gulp": "^3.9.0",
+    "gulp": "^3.9.1",
     "gulp-awspublish": "^3.0.1",
     "gulp-connect": "^2.2.0",
     "gulp-less": "^3.0.3",
